requirements.txt
paho-mqtt
python-dateutil
zahlwort2num